fix(CB2-22294): move shared logic into service
